• transferfunctions.py   remove ExponentialDecay and AcceleratingDecay
• integratorfunctions.py
  add ExponentialDecay and AcceleratingDecay
